+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Zipkin tags used for InstrumentationLibrary changed from library.name,
+  library.version to otel.library.name, otel.library.version respectively.
+
 ## 0.8.0-beta.1
 
 Released 2020-Nov-5